backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==3.2.16
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
drf-yasg==1.21.4
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.10.0
django-cors-headers==3.13.0